Перевод всех Entity/Domain ID с String на Int, UnauthorizedActionSnackbar
- Все Room entities, domain модели, DAOs, репозитории, ViewModels, навигация: id String → Int
- Мапперы DTO→Domain/Entity: добавлен .toInt() для id полей
- Repository implementations: .toString() при вызовах remote data sources и конструировании DTO body
- DTO не тронуты, remote data sources не тронуты
- Обновлены UI компоненты: callback типы, scoreboard views, preview data
- UnauthorizedActionSnackbar: доработка

+++ composeApp/src/wasmJsMain/kotlin/com/bashkevich/tennisscorekeeper/main.kt
@@ -26,7 +26,7 @@ fun main() {
             when {
                 // Identifies the corresponding route and navigates to it
                 initRoute.contains("scoreboard") -> {
-                    val matchId = initRoute.substring(findNthOccurrence(initRoute,'/',1)+1,findNthOccurrence(initRoute,'/',2))
+                    val matchId = initRoute.substring(findNthOccurrence(initRoute,'/',1)+1,findNthOccurrence(initRoute,'/',2)).toInt()
                     navController.navigate(ScoreboardRoute(matchId = matchId))
                 }
                 // Identifies the corresponding route and navigates to it
@@ -37,7 +37,7 @@ fun main() {
                     navController.navigate(AddTournamentRoute)
                 }
                 initRoute.matches(Regex("tournaments/[0-9]+/addMatch")) ->{
-                    val tournamentId = initRoute.substring(findNthOccurrence(initRoute,'/',1)+1,findNthOccurrence(initRoute,'/',2))
+                    val tournamentId = initRoute.substring(findNthOccurrence(initRoute,'/',1)+1,findNthOccurrence(initRoute,'/',2)).toInt()
 
                     navController.navigate(AddMatchRoute(tournamentId = tournamentId))
                 }
